move /types to src/types
